wip: wrap react-router routes for compatibility
