+++ packages/react-router/src/index.tsx
@@ -16,7 +16,6 @@ export * from "react-router-dom";
 
 export { Link } from "./Link";
 export type { LinkProps } from "./Link";
-export { Route } from "./Route";
 export type { RouteProps } from "./Route";
 export { Prompt } from "./Prompt";
 export type { PromptProps } from "./Prompt";
